week 7 : final version

+++ week 7/cv_lab05_code/visualizeSegmentationResults.m
@@ -20,6 +20,6 @@ for d=1:D
 end
 
 figure;
-imshow(valImg);
+imshow(uint8(valImg));
 
 end
